waitp-1364-split-up-types-file
