+++ packages/tupaia-web/src/types/index.d.ts
@@ -1,2 +1,4 @@
 export * from './types';
+export * from './report';
+export * from './dashboardItem';
 export * from './helpers';
